+++ packages/common/src/types/dbt.ts
@@ -1,6 +1,16 @@
 import { DepGraph } from 'dependency-graph';
 import { DbtError, ParseError } from './errors';
-import { DimensionType, MetricType } from './field';
+import {
+    defaultSql,
+    DimensionType,
+    FieldType,
+    friendlyName,
+    Metric,
+    MetricType,
+    Source,
+} from './field';
+import { AdditionalMetric } from './metricQuery';
+import { TableBase } from './table';
 
 export enum SupportedDbtAdapter {
     BIGQUERY = 'bigquery',
@@ -159,3 +169,223 @@ export const buildModelGraph = (
     });
     return depGraph;
 };
+
+export interface DbtCatalogNode {
+    metadata: DbtCatalogNodeMetadata;
+    columns: {
+        [k: string]: DbtCatalogNodeColumn;
+    };
+}
+
+export interface DbtCatalogNodeMetadata {
+    type: string;
+    database: string | null;
+    schema: string;
+    name: string;
+    comment?: string;
+    owner?: string;
+}
+
+export interface DbtCatalogNodeColumn {
+    type: string;
+    comment?: string;
+    index: number;
+    name: string;
+}
+
+export interface DbtRpcDocsGenerateResults {
+    nodes: {
+        [k: string]: DbtCatalogNode;
+    };
+}
+
+export const isDbtRpcDocsGenerateResults = (
+    results: Record<string, any>,
+): results is DbtRpcDocsGenerateResults =>
+    'nodes' in results &&
+    typeof results.nodes === 'object' &&
+    results.nodes !== null &&
+    Object.values(results.nodes).every(
+        (node) =>
+            typeof node === 'object' &&
+            node !== null &&
+            'metadata' in node &&
+            'columns' in node,
+    );
+
+export interface DbtPackage {
+    package: string;
+    version: string;
+}
+
+export interface DbtPackages {
+    packages: DbtPackage[];
+}
+
+export const isDbtPackages = (
+    results: Record<string, any>,
+): results is DbtPackages => 'packages' in results;
+type DbtMetricFilter = {
+    field: string;
+    operator: string;
+    value: string;
+};
+export type DbtMetric = {
+    unique_id: string;
+    package_name: string;
+    path: string;
+    root_path: string;
+    original_file_path: string;
+    model: string;
+    name: string;
+    description: string;
+    label: string;
+    type: string;
+    timestamp: string | null;
+    filters: DbtMetricFilter[];
+    time_grains: string[];
+    dimensions: string[];
+    resource_type?: 'metric';
+    meta?: Record<string, any> & DbtMetricLightdashMetadata;
+    tags?: string[];
+    sql?: string | null;
+};
+export type DbtMetricLightdashMetadata = {
+    hidden?: boolean;
+};
+
+export interface DbtManifest {
+    nodes: Record<string, DbtNode>;
+    metadata: DbtRawManifestMetadata;
+    metrics: Record<string, DbtMetric>;
+}
+
+export interface DbtRawManifestMetadata {
+    dbt_schema_version: string;
+    generated_at: string;
+    adapter_type: string;
+}
+
+export interface DbtManifestMetadata extends DbtRawManifestMetadata {
+    adapter_type: SupportedDbtAdapter;
+}
+
+const isDbtRawManifestMetadata = (x: any): x is DbtRawManifestMetadata =>
+    typeof x === 'object' &&
+    x !== null &&
+    'dbt_schema_version' in x &&
+    'generated_at' in x &&
+    'adapter_type' in x;
+export const isSupportedDbtAdapter = (
+    x: DbtRawManifestMetadata,
+): x is DbtManifestMetadata =>
+    isDbtRawManifestMetadata(x) &&
+    Object.values<string>(SupportedDbtAdapter).includes(x.adapter_type);
+
+export interface DbtRpcGetManifestResults {
+    manifest: DbtManifest;
+}
+
+export const isDbtRpcManifestResults = (
+    results: Record<string, any>,
+): results is DbtRpcGetManifestResults =>
+    'manifest' in results &&
+    typeof results.manifest === 'object' &&
+    results.manifest !== null &&
+    'nodes' in results.manifest &&
+    'metadata' in results.manifest &&
+    'metrics' in results.manifest &&
+    isDbtRawManifestMetadata(results.manifest.metadata);
+
+export interface DbtRpcCompileResults {
+    results: { node: DbtNode }[];
+}
+
+export const isDbtRpcCompileResults = (
+    results: Record<string, any>,
+): results is DbtRpcCompileResults =>
+    'results' in results &&
+    Array.isArray(results.results) &&
+    results.results.every(
+        (result) =>
+            typeof result === 'object' &&
+            result !== null &&
+            'node' in result &&
+            typeof result.node === 'object' &&
+            result.node !== null &&
+            'unique_id' in result.node &&
+            'resource_type' in result.node,
+    );
+
+export interface DbtRpcRunSqlResults {
+    results: {
+        table: { column_names: string[]; rows: any[][] };
+    }[];
+}
+
+export const isDbtRpcRunSqlResults = (
+    results: Record<string, any>,
+): results is DbtRpcRunSqlResults =>
+    'results' in results &&
+    Array.isArray(results.results) &&
+    results.results.every(
+        (result) =>
+            typeof result === 'object' &&
+            result !== null &&
+            'table' in result &&
+            typeof result.table === 'object' &&
+            result.table !== null &&
+            'column_names' in result.table &&
+            Array.isArray(result.table.column_names) &&
+            'rows' in result.table &&
+            Array.isArray(result.table.rows),
+    );
+type ConvertMetricArgs = {
+    modelName: string;
+    columnName: string;
+    name: string;
+    metric: DbtColumnLightdashMetric;
+    source?: Source;
+    tableLabel: string;
+};
+export const convertMetric = ({
+    modelName,
+    columnName,
+    name,
+    metric,
+    source,
+    tableLabel,
+}: ConvertMetricArgs): Metric => ({
+    fieldType: FieldType.METRIC,
+    name,
+    label: metric.label || friendlyName(name),
+    sql: metric.sql || defaultSql(columnName),
+    table: modelName,
+    tableLabel,
+    type: metric.type,
+    isAutoGenerated: false,
+    description:
+        metric.description ||
+        `${friendlyName(metric.type)} of ${friendlyName(columnName)}`,
+    source,
+    hidden: !!metric.hidden,
+    round: metric.round,
+    format: metric.format,
+});
+type ConvertAdditionalMetricArgs = {
+    additionalMetric: AdditionalMetric;
+    table: TableBase;
+    dimension?: string;
+};
+export const convertAdditionalMetric = ({
+    additionalMetric,
+    table,
+    dimension,
+}: ConvertAdditionalMetricArgs): Metric =>
+    convertMetric({
+        modelName: table.name,
+        columnName: dimension || '',
+        name: additionalMetric.name,
+        metric: additionalMetric,
+        tableLabel: table.label,
+    });