+++ packages/common/src/types/table.ts
@@ -1,0 +1,8 @@
+export type TableBase = {
+    name: string; // Must be sql friendly (a-Z, 0-9, _)
+    label: string; // Friendly name
+    description?: string; // Optional description of table
+    database: string;
+    schema: string;
+    sqlTable: string; // The sql identifier for the table
+};